improved PopSizes to not rely on results
